backend/requirements.txt
fastapi
uvicorn
python-dotenv
alpaca-trade-api
pandas
numpy
